IPython console: Move ssh utilities from its plugin to utils/ssh.py
