+++ spyder/plugins/ipythonconsole/plugin.py
@@ -14,7 +14,6 @@ IPython Console plugin based on QtConsole
 # pylint: disable=R0201
 
 # Standard library imports
-import atexit
 import os
 import os.path as osp
 import uuid
@@ -31,8 +30,6 @@ from qtpy.QtWidgets import (QApplication, QHBoxLayout, QMessageBox,
                             QVBoxLayout, QWidget)
 from traitlets.config.loader import Config, load_pyconfig_files
 from zmq.ssh import tunnel as zmqtunnel
-if not os.name == 'nt':
-    import pexpect
 
 # Local imports
 from spyder import dependencies
@@ -43,6 +40,7 @@ from spyder.py3compat import is_string, PY2, to_text_string
 from spyder.plugins.ipythonconsole.confpage import IPythonConsoleConfigPage
 from spyder.plugins.ipythonconsole.utils.kernelspec import SpyderKernelSpec
 from spyder.plugins.ipythonconsole.utils.style import create_qss_style
+from spyder.plugins.ipythonconsole.utils.ssh import openssh_tunnel
 from spyder.utils.qthelpers import create_action, MENU_SEPARATOR
 from spyder.utils import icon_manager as ima
 from spyder.utils import encoding, programs, sourcecode
@@ -75,91 +73,7 @@ MATPLOTLIB_REQVER = '>=2.0.0'
 dependencies.add("matplotlib", _("Display 2D graphics in the IPython Console"),
                  required_version=MATPLOTLIB_REQVER, optional=True)
 
-#------------------------------------------------------------------------------
-# Existing kernels
-#------------------------------------------------------------------------------
-# Replacing pyzmq openssh_tunnel method to work around the issue
-# https://github.com/zeromq/pyzmq/issues/589 which was solved in pyzmq
-# https://github.com/zeromq/pyzmq/pull/615
-def _stop_tunnel(cmd):
-    pexpect.run(cmd)
 
-def openssh_tunnel(self, lport, rport, server, remoteip='127.0.0.1',
-                   keyfile=None, password=None, timeout=0.4):
-    ssh="ssh "
-    if keyfile:
-        ssh += "-i " + keyfile
-    
-    if ':' in server:
-        server, port = server.split(':')
-        ssh += " -p %s" % port
-    
-    cmd = "%s -O check %s" % (ssh, server)
-    (output, exitstatus) = pexpect.run(cmd, withexitstatus=True)
-    if not exitstatus:
-        pid = int(output[output.find("(pid=")+5:output.find(")")]) 
-        cmd = "%s -O forward -L 127.0.0.1:%i:%s:%i %s" % (
-            ssh, lport, remoteip, rport, server)
-        (output, exitstatus) = pexpect.run(cmd, withexitstatus=True)
-        if not exitstatus:
-            atexit.register(_stop_tunnel, cmd.replace("-O forward",
-                                                      "-O cancel",
-                                                      1))
-            return pid
-    cmd = "%s -f -S none -L 127.0.0.1:%i:%s:%i %s sleep %i" % (
-                                  ssh, lport, remoteip, rport, server, timeout)
-    
-    # pop SSH_ASKPASS from env
-    env = os.environ.copy()
-    env.pop('SSH_ASKPASS', None)
-    
-    ssh_newkey = 'Are you sure you want to continue connecting'
-    tunnel = pexpect.spawn(cmd, env=env)
-    failed = False
-    while True:
-        try:
-            i = tunnel.expect([ssh_newkey, '[Pp]assword:'], timeout=.1)
-            if i==0:
-                host = server.split('@')[-1]
-                question = _("The authenticity of host <b>%s</b> can't be "
-                             "established. Are you sure you want to continue "
-                             "connecting?") % host
-                reply = QMessageBox.question(self, _('Warning'), question,
-                                             QMessageBox.Yes | QMessageBox.No,
-                                             QMessageBox.No)
-                if reply == QMessageBox.Yes:
-                    tunnel.sendline('yes')
-                    continue
-                else:
-                    tunnel.sendline('no')
-                    raise RuntimeError(
-                       _("The authenticity of the host can't be established"))
-            if i==1 and password is not None:
-                tunnel.sendline(password) 
-        except pexpect.TIMEOUT:
-            continue
-        except pexpect.EOF:
-            if tunnel.exitstatus:
-                raise RuntimeError(_("Tunnel '%s' failed to start") % cmd)
-            else:
-                return tunnel.pid
-        else:
-            if failed or password is None:
-                raise RuntimeError(_("Could not connect to remote host"))
-                # TODO: Use this block when pyzmq bug #620 is fixed
-                # # Prompt a passphrase dialog to the user for a second attempt
-                # password, ok = QInputDialog.getText(self, _('Password'),
-                #             _('Enter password for: ') + server,
-                #             echo=QLineEdit.Password)
-                # if ok is False:
-                #      raise RuntimeError('Could not connect to remote host.') 
-            tunnel.sendline(password)
-            failed = True
-
-
-#------------------------------------------------------------------------------
-# Plugin widget
-#------------------------------------------------------------------------------
 class IPythonConsole(SpyderPluginWidget):
     """
     IPython Console plugin
